move ResetForm to page component
